+++ egui/src/context.rs
@@ -583,13 +583,14 @@ impl Context {
     }
 
     /// Tessellate the given shapes into triangle meshes.
-    pub fn tessellate(&self, shapes: Vec<ClippedShape>) -> PaintJobs {
+    pub fn tessellate(&self, shapes: Vec<ClippedShape>) -> Vec<ClippedMesh> {
         let mut tessellation_options = self.memory().options.tessellation_options;
         tessellation_options.aa_size = 1.0 / self.pixels_per_point();
         let paint_stats = PaintStats::from_shapes(&shapes); // TODO: internal allocations
-        let paint_jobs = tessellator::tessellate_shapes(shapes, tessellation_options, self.fonts());
-        *self.paint_stats.lock() = paint_stats.with_paint_jobs(&paint_jobs);
-        paint_jobs
+        let clipped_meshes =
+            tessellator::tessellate_shapes(shapes, tessellation_options, self.fonts());
+        *self.paint_stats.lock() = paint_stats.with_clipped_meshes(&clipped_meshes);
+        clipped_meshes
     }
 
     // ---------------------------------------------------------------------